feat: revamp vault page
