+++ src/components/ContractWriteResultCard/ContractWriteResultCard_View.tsx
@@ -7,13 +7,13 @@ const ContractWriteResultCard = ({ hash, isSuccess, errorMsg }: Props) => {
   return (
     <>
       {isSuccess && (
-        <div className="mt-5 w-96 px-10 py-5 rounded-md shadow  bg-green-300 text-green-800  break-all">
+        <div className="px-10 py-5 rounded-md shadow  bg-green-300 text-green-800  break-all">
           Success <br />
           Transaction Hash : {hash}
         </div>
       )}
       {errorMsg && (
-        <div className="mt-5 w-96 px-10 py-5 rounded-md shadow  bg-red-300  text-red-800 break-words">
+        <div className="px-10 py-5 rounded-md shadow  bg-red-300  text-red-800 break-words">
           Error <br />
           {errorMsg}
         </div>
