+++ src/pages/Vault/VaultDetail/VaultDetail_View.tsx
@@ -1,0 +1,37 @@
+type Props = {
+  tokenSymbol: string;
+  vaultAddress: string;
+  vaultBalance: string;
+  userBalance: string;
+  performance: string;
+  asset: string;
+};
+
+const VaultDetailView = ({
+  tokenSymbol,
+  vaultAddress,
+  vaultBalance,
+  userBalance,
+  performance,
+  asset
+}: Props) => {
+  return (
+    <div className="p-5 w-64">
+      <h1 className="text-3xl mb-2">{tokenSymbol} Vault</h1>
+      <div>Share: {userBalance ?? "connect wallet to see your balance"}</div>
+      <div className="break-all">
+        Vault Address: <br />
+        {vaultAddress}
+      </div>
+      <div className="break-all">
+        Underlying <br />
+        {asset}
+      </div>
+      <div>TLV: {vaultBalance}</div>
+
+      <div>APY: {performance}%</div>
+    </div>
+  );
+};
+
+export default VaultDetailView;
